Moving TestClient to ext-pgw lib
