+++ ext-pgw/build.gradle
@@ -1,0 +1,22 @@
+plugins {
+    id "java-library"
+    id "idea"
+}
+
+dependencies {
+
+    implementation project(':ocsgw')
+
+    implementation ('org.mobicents.diameter:jdiameter-api:1.7.1-123') {
+        exclude module: 'netty-all'
+    }
+    implementation ('org.mobicents.diameter:jdiameter-impl:1.7.1-123') {
+        exclude module: 'netty-all'
+    }
+
+    testCompile 'org.junit.jupiter:junit-jupiter-api:5.0.3'
+    testRuntime 'org.junit.jupiter:junit-jupiter-engine:5.0.3'
+
+    testCompile 'junit:junit:4.12'
+    testRuntime 'org.junit.vintage:junit-vintage-engine:4.12.3'
+}
