+++ ext-pgw/src/main/java/com/telenordigital/ext_pgw/TestClient.java
@@ -1,33 +1,17 @@
+package com.telenordigital.ext_pgw;
 
-package com.telenordigital.ocsgw.testclient;
+import org.apache.log4j.Logger;
+import org.apache.log4j.PropertyConfigurator;
+import org.jdiameter.api.*;
+import org.jdiameter.api.cca.events.JCreditControlRequest;
+import org.jdiameter.server.impl.StackImpl;
+import org.jdiameter.server.impl.helpers.XMLConfiguration;
 
 import java.io.IOException;
 import java.io.InputStream;
 import java.util.Properties;
 import java.util.Set;
 import java.util.concurrent.TimeUnit;
-
-import org.apache.log4j.Logger;
-import org.apache.log4j.PropertyConfigurator;
-import org.jdiameter.api.Answer;
-import org.jdiameter.api.ApplicationId;
-import org.jdiameter.api.Avp;
-import org.jdiameter.api.AvpSet;
-import org.jdiameter.api.Configuration;
-import org.jdiameter.api.EventListener;
-import org.jdiameter.api.IllegalDiameterStateException;
-import org.jdiameter.api.InternalException;
-import org.jdiameter.api.Message;
-import org.jdiameter.api.Network;
-import org.jdiameter.api.OverloadException;
-import org.jdiameter.api.Request;
-import org.jdiameter.api.RouteException;
-import org.jdiameter.api.Session;
-import org.jdiameter.api.SessionFactory;
-import org.jdiameter.api.Stack;
-import org.jdiameter.api.cca.events.JCreditControlRequest;
-import org.jdiameter.server.impl.StackImpl;
-import org.jdiameter.server.impl.helpers.XMLConfiguration;
 
 public class TestClient implements EventListener<Request, Answer> {
