+++ ext-pgw/src/test/resources/config.properties
@@ -1,0 +1,7 @@
+# Type of data storage (Local or gRPC)
+DataStoreType=gRPC
+
+# The ip:port on which the gRPC service will be receiving incoming connections.
+GrpcServer=172.16.238.4:8082
+
+GrpcEnryption=false
