+++ ext-pgw/src/test/resources/log4j.properties
@@ -1,0 +1,15 @@
+# Set root logger level to DEBUG and its only appender to A1.
+log4j.rootLogger=DEBUG, A1
+
+# A1 is set to be a ConsoleAppender.
+log4j.appender.A1=org.apache.log4j.ConsoleAppender
+
+# A1 uses PatternLayout.
+log4j.appender.A1.layout=org.apache.log4j.PatternLayout
+log4j.appender.A1.layout.ConversionPattern=%d{dd MMM yyyy HH:mm:ss,SSS} %-4r [%t] %-5p %c %x - %m%n
+
+
+log4j.logger.org.mobicents.tests.diameter=INFO
+log4j.logger.org.jdiameter=INFO
+log4j.logger.jdiameter=OFF
+log4j.logger.org.jdiameter.common.impl.validation=OFF
