+++ settings.gradle
@@ -4,8 +4,10 @@ include ':ocs-api'
 include ':prime'
 include ':ocsgw'
 include ':ostelco-lib'
+include ':ext-pgw'
 
 project(':ocs-api').projectDir = "$rootDir/ocs-api" as File
 project(':prime').projectDir = "$rootDir/prime" as File
 project(':ocsgw').projectDir = "$rootDir/ocsgw" as File
 project(':ostelco-lib').projectDir = "$rootDir/ostelco-lib" as File
+project(':ext-pgw').projectDir = "$rootDir/ext-pgw" as File
